fleshed out tests for GetAllOwnedTokens
