+++ Source/nt.Server/Features/WebThree/Contracts/Herc1155/GetAllOwnedTokens/GetAllOwnedTokensFeaturesHandler.cs
@@ -1,12 +1,11 @@
 ﻿namespace nt.Server.Features.WebThree.Contracts.Herc1155.GetAllOwnedTokens
 {
   using MediatR;
-  using nt.Shared.Features.WebThree.Contracts.Herc1155;
-  using nt.Server.Services.WebThree.Contracts.Herc1155.Functions;
+  using nt.Shared.Features.WebThree.Contracts.Herc1155.GetAllOwnedTokens;
+  using nt.Server.Services.WebThree.Contracts.Herc1155.GetAllOwnedTokens;
   using System.Threading;
   using System.Threading.Tasks;
   using nt.Shared.Constants.AccountAddresses;
-
 
   public class GetAllOwnedTokensServerFeaturesHandler : IRequestHandler<GetAllOwnedTokensSharedRequest, GetAllOwnedTokensSharedResponse>
   {
