+++ Source/nt.Server/Features/WebThree/Contracts/Herc1155/GetAllOwnedTokens/GetAllOwnedTokensServerFeatureController.cs
@@ -3,7 +3,7 @@
   using System.Threading.Tasks;
   using nt.Server.Features.Base;
   using Microsoft.AspNetCore.Mvc;
-  using nt.Shared.Features.WebThree.Contracts.Herc1155;
+  using nt.Shared.Features.WebThree.Contracts.Herc1155.GetAllOwnedTokens;
 
   [Route(GetAllOwnedTokensSharedRequest.Route)]
   public class GetAllOwnedTokensServerFeaturesController : BaseController<GetAllOwnedTokensSharedRequest, GetAllOwnedTokensSharedResponse>
