+++ Source/nt.Server/Services/WebThree/Contracts/Herc1155/Functions/GetAllOwnedTokens/GetAllOwnedTokensServiceRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace nt.Server.Services.WebThree.Contracts.Herc1155
+﻿namespace nt.Server.Services.WebThree.Contracts.Herc1155.GetAllOwnedTokens
 {
   using MediatR;
