+++ Source/nt.Shared/Features/WebThree/Contracts/Herc1155/GetAllOwnedTokens/GetAllOwnedTokensFunctionInput.cs
@@ -1,4 +1,4 @@
-﻿namespace nt.Shared.Features.WebThree.Contracts.Herc1155
+﻿namespace nt.Shared.Features.WebThree.Contracts.Herc1155.GetAllOwnedTokens
 {
   using Nethereum.ABI.FunctionEncoding.Attributes;
   using Nethereum.Web3.Accounts;
@@ -9,5 +9,6 @@
   {
     [Parameter(type: "address", name: "owner", order: 1)]
     public string TokenOwner { get; set; } // this will be the default Edge Eth Address
+    
   }
 }
