+++ Source/nt.Shared/Features/WebThree/Contracts/Herc1155/GetAllOwnedTokens/GetAllOwnedTokensSharedResponse.cs
@@ -1,4 +1,4 @@
-namespace nt.Shared.Features.WebThree.Contracts.Herc1155
+namespace nt.Shared.Features.WebThree.Contracts.Herc1155.GetAllOwnedTokens
 {
   using System;
   using System.Collections.Generic;
